requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
scipy>=1.13
matplotlib>=3.8
joblib>=1.4
tqdm>=4.66
emoji>=2.12.1
unicodedata2>=15.1
beautifulsoup4>=4.12
lxml>=5.2
html2text>=2024.2.26
imapclient>=2.3.1
python-dotenv>=1.0.1
